refactor(Filters): Refactor filters into a more global file, and add permission_filter.
